Added Product class

+++ Bill.java
@@ -31,7 +31,7 @@ public class Bill{
 
 	public long calculateTotalCost(){
 
-		int totalDiscounts = 0;
+		long totalDiscounts = 0;
 
 		for(int i = 0; i < discounts.size(); i++)
